Conflicts:
	recovery-webservice/src/main/java/es/pfsgroup/plugin/rem/api/impl/UvemManager.java

Former-commit-id: 5ca036a2a02a9d3ee03fcb6a025f8f2fc8b6a5db

+++ recovery-webservice/src/main/java/es/pfsgroup/plugin/rem/activo/ActivoManager.java
@@ -39,14 +39,17 @@ import es.pfsgroup.framework.paradise.utils.BeanUtilNotNull;
 import es.pfsgroup.framework.paradise.utils.DtoPage;
 import es.pfsgroup.plugin.recovery.coreextension.utils.api.UtilDiccionarioApi;
 import es.pfsgroup.plugin.recovery.nuevoModeloBienes.model.DDUnidadPoblacional;
+import es.pfsgroup.plugin.recovery.nuevoModeloBienes.model.NMBValoracionesBien;
 import es.pfsgroup.plugin.rem.activo.dao.ActivoDao;
 import es.pfsgroup.plugin.rem.adapter.ActivoAdapter;
 import es.pfsgroup.plugin.rem.adapter.GenericAdapter;
 import es.pfsgroup.plugin.rem.api.ActivoApi;
 import es.pfsgroup.plugin.rem.api.TrabajoApi;
+import es.pfsgroup.plugin.rem.api.UvemManagerApi;
 import es.pfsgroup.plugin.rem.factory.TabActivoFactoryApi;
 import es.pfsgroup.plugin.rem.model.Activo;
 import es.pfsgroup.plugin.rem.model.ActivoAdjuntoActivo;
+import es.pfsgroup.plugin.rem.model.ActivoAgrupacionActivo;
 import es.pfsgroup.plugin.rem.model.ActivoBancario;
 import es.pfsgroup.plugin.rem.model.ActivoCatastro;
 import es.pfsgroup.plugin.rem.model.ActivoCondicionEspecifica;
@@ -57,7 +60,9 @@ import es.pfsgroup.plugin.rem.model.ActivoHistoricoValoraciones;
 import es.pfsgroup.plugin.rem.model.ActivoInformeComercialHistoricoMediador;
 import es.pfsgroup.plugin.rem.model.ActivoOferta;
 import es.pfsgroup.plugin.rem.model.ActivoPropietarioActivo;
+import es.pfsgroup.plugin.rem.model.ActivoProveedor;
 import es.pfsgroup.plugin.rem.model.ActivoSituacionPosesoria;
+import es.pfsgroup.plugin.rem.model.ActivoTasacion;
 import es.pfsgroup.plugin.rem.model.ActivoValoraciones;
 import es.pfsgroup.plugin.rem.model.DtoActivoDatosRegistrales;
 import es.pfsgroup.plugin.rem.model.DtoActivoFichaCabecera;
@@ -77,6 +82,7 @@ import es.pfsgroup.plugin.rem.model.DtoOfertaActivo;
 import es.pfsgroup.plugin.rem.model.DtoPrecioVigente;
 import es.pfsgroup.plugin.rem.model.DtoPropuestaActivosVinculados;
 import es.pfsgroup.plugin.rem.model.DtoPropuestaFilter;
+import es.pfsgroup.plugin.rem.model.DtoTasacion;
 import es.pfsgroup.plugin.rem.model.ExpedienteComercial;
 import es.pfsgroup.plugin.rem.model.GastosExpediente;
 import es.pfsgroup.plugin.rem.model.Oferta;
@@ -93,8 +99,11 @@ import es.pfsgroup.plugin.rem.model.dd.DDEstadoOferta;
 import es.pfsgroup.plugin.rem.model.dd.DDEstadoPublicacion;
 import es.pfsgroup.plugin.rem.model.dd.DDEstadosExpedienteComercial;
 import es.pfsgroup.plugin.rem.model.dd.DDEstadosVisitaOferta;
+import es.pfsgroup.plugin.rem.model.dd.DDMotivoComercializacion;
 import es.pfsgroup.plugin.rem.model.dd.DDSituacionComercial;
+import es.pfsgroup.plugin.rem.model.dd.DDSubcartera;
 import es.pfsgroup.plugin.rem.model.dd.DDSubtipoTrabajo;
+import es.pfsgroup.plugin.rem.model.dd.DDTipoAgrupacion;
 import es.pfsgroup.plugin.rem.model.dd.DDTipoDocumentoActivo;
 import es.pfsgroup.plugin.rem.model.dd.DDTipoFoto;
 import es.pfsgroup.plugin.rem.model.dd.DDTipoPrecio;
@@ -106,9 +115,6 @@ import es.pfsgroup.plugin.rem.visita.dao.VisitaDao;
 
 @Service("activoManager")
 public class ActivoManager extends BusinessOperationOverrider<ActivoApi> implements ActivoApi {
-
-	private final String ESTADO_PORTALES_EXTERNOS_PUBLICADO = "Ha estado publicado";
-	private final String ESTADO_PORTALES_EXTERNOS_NO_PUBLICADO = "No ha estado publicado";
 	
 	SimpleDateFormat ft = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss");
 
@@ -143,6 +149,9 @@ public class ActivoManager extends BusinessOperationOverrider<ActivoApi> impleme
 
 	@Autowired
 	private VisitaDao visitasDao;
+	
+	@Autowired
+	private UvemManagerApi uvemManagerApi;
 
 	@Override
 	public String managerName() {
@@ -437,12 +446,29 @@ public class ActivoManager extends BusinessOperationOverrider<ActivoApi> impleme
 	@Override
 	@Transactional(readOnly = false)
 	public boolean deleteValoracionPrecio(Long id) {
-
+		
+		return deleteValoracionPrecioConGuardadoEnHistorico(id,true);
+	}
+	
+	@Override
+	@Transactional(readOnly = false)
+	public boolean deleteValoracionPrecioConGuardadoEnHistorico(Long id, Boolean guardadoEnHistorico) {
+		
 		Filter filtro = genericDao.createFilter(FilterType.EQUALS, "id", id);
 		ActivoValoraciones activoValoracion = genericDao.get(ActivoValoraciones.class, filtro);
-		saveActivoValoracionHistorico(activoValoracion);
-		// genericDao.deleteById(ActivoValoraciones.class, id);
-		activoDao.deleteValoracionById(id);
+		
+		if(guardadoEnHistorico) {
+			saveActivoValoracionHistorico(activoValoracion);
+			activoDao.deleteValoracionById(id);
+		}
+		else if(!Checks.esNulo(activoValoracion.getGestor()) && !adapter.getUsuarioLogado().equals(activoValoracion.getGestor())) {
+			//Si el usuario logado es distinto al que ha creado la valoracion, no puede borrarla sin historico
+			return false;
+		}
+		else {
+			//Al anular el precio vigente, se hace un borrado lógico, y no se inserta en el histórico
+			genericDao.deleteById(ActivoValoraciones.class, id);
+		}
 
 		return true;
 	}
@@ -1020,7 +1046,7 @@ public class ActivoManager extends BusinessOperationOverrider<ActivoApi> impleme
 	@Override
 	public List<DtoHistoricoMediador> getHistoricoMediadorByActivo(Long idActivo){
 		Filter filtro = genericDao.createFilter(FilterType.EQUALS, "activo.id", idActivo);
-		Order order = new Order(OrderType.ASC, "id");
+		Order order = new Order(OrderType.DESC, "id");
 		List<ActivoInformeComercialHistoricoMediador> listaHistoricoMediador = genericDao
 				.getListOrdered(ActivoInformeComercialHistoricoMediador.class, order, filtro);
 		
@@ -1029,7 +1055,8 @@ public class ActivoManager extends BusinessOperationOverrider<ActivoApi> impleme
 		for (ActivoInformeComercialHistoricoMediador historico : listaHistoricoMediador) {
 			DtoHistoricoMediador dtoHistoricoMediador = new DtoHistoricoMediador();
 			try {
-				beanUtilNotNull.copyProperty(dtoHistoricoMediador, "id", idActivo);
+				beanUtilNotNull.copyProperty(dtoHistoricoMediador, "id", historico.getId());
+				beanUtilNotNull.copyProperty(dtoHistoricoMediador, "idActivo", idActivo);
 				beanUtilNotNull.copyProperty(dtoHistoricoMediador, "fechaDesde", historico.getFechaDesde());
 				beanUtilNotNull.copyProperty(dtoHistoricoMediador, "fechaHasta", historico.getFechaHasta());
 				if(!Checks.esNulo(historico.getMediadorInforme())){
@@ -1048,6 +1075,52 @@ public class ActivoManager extends BusinessOperationOverrider<ActivoApi> impleme
 		}
 		
 		return listaDtoHistoricoMediador;
+	}
+	
+	@Override
+	@Transactional(readOnly = false)
+	public Boolean createHistoricoMediador(DtoHistoricoMediador dto) {
+		ActivoInformeComercialHistoricoMediador historicoMediador = new ActivoInformeComercialHistoricoMediador();
+		Activo activo = null;
+		
+		if(!Checks.esNulo(dto.getIdActivo())) {
+			activo = activoDao.get(dto.getIdActivo());
+		}
+		
+		try {
+			// Terminar periodo de vigencia del último proveedor (fecha hasta).
+			if(!Checks.esNulo(activo)) {
+				Filter activoIDFiltro = genericDao.createFilter(FilterType.EQUALS, "activo.id", activo.getId());
+				Order order = new Order(OrderType.DESC, "id");
+				List<ActivoInformeComercialHistoricoMediador> historicoMediadorlist = genericDao.getListOrdered(ActivoInformeComercialHistoricoMediador.class, order, activoIDFiltro);
+				if(!Checks.estaVacio(historicoMediadorlist)) {
+					ActivoInformeComercialHistoricoMediador historicoAnteriorMediador = historicoMediadorlist.get(0); // El primero es el de ID más alto (el último).
+					beanUtilNotNull.copyProperty(historicoAnteriorMediador, "fechaHasta", new Date());
+					genericDao.save(ActivoInformeComercialHistoricoMediador.class, historicoAnteriorMediador);
+				}
+			}
+
+			// Generar la nueva entrada de HistoricoMediador.
+			beanUtilNotNull.copyProperty(historicoMediador, "fechaDesde", new Date());
+			beanUtilNotNull.copyProperty(historicoMediador, "activo", activo);
+			
+			if(!Checks.esNulo(dto.getMediador())) {
+				Filter proveedorFiltro = genericDao.createFilter(FilterType.EQUALS, "id", Long.parseLong(dto.getMediador()));
+				ActivoProveedor proveedor = genericDao.get(ActivoProveedor.class, proveedorFiltro);
+				beanUtilNotNull.copyProperty(historicoMediador, "mediadorInforme", proveedor);
+			}
+
+			genericDao.save(ActivoInformeComercialHistoricoMediador.class, historicoMediador);
+			
+		} catch (IllegalAccessException e) {
+			e.printStackTrace();
+			return false;
+		} catch (InvocationTargetException e) {
+			e.printStackTrace();
+			return false;
+		}
+		
+		return true;
 	}
 
 	@Override
@@ -1086,7 +1159,7 @@ public class ActivoManager extends BusinessOperationOverrider<ActivoApi> impleme
 	@Override
 	public DtoDatosPublicacion getDatosPublicacionByActivo(Long idActivo) {
 		// Obtener los estados y sumar los dias de cada fase aplicando criterio de funcional además comprobar
-		// si alguno de ellos es de publicación/publicación forzada y especificar estado publicación en portales externos.
+		// si alguno de ellos es de publicación/publicación forzada.
 		Filter filtro = genericDao.createFilter(FilterType.EQUALS, "activo.id", idActivo);
 		Order order = new Order(OrderType.ASC, "id");
 		List<ActivoHistoricoEstadoPublicacion> listaEstadosPublicacion = genericDao
@@ -1094,16 +1167,9 @@ public class ActivoManager extends BusinessOperationOverrider<ActivoApi> impleme
 		
 		int dias = 0;
 		boolean despublicado = false;
-		boolean estadoPublicacionPortalesExternos = false;
 		
 		for (ActivoHistoricoEstadoPublicacion estado : listaEstadosPublicacion) {
 			if(!Checks.esNulo(estado.getEstadoPublicacion())){
-				if(DDEstadoPublicacion.CODIGO_PUBLICADO.equals(estado.getEstadoPublicacion().getCodigo()) || 
-						DDEstadoPublicacion.CODIGO_PUBLICADO_FORZADO.equals(estado.getEstadoPublicacion().getCodigo())){
-					// Comprobar si ha estado publicado en algún estado para especificarlo en estado publicación portales externos.
-					estadoPublicacionPortalesExternos = true;
-				}
-				
 				if(despublicado && (DDEstadoPublicacion.CODIGO_PUBLICADO.equals(estado.getEstadoPublicacion().getCodigo()) || 
 						DDEstadoPublicacion.CODIGO_PUBLICADO_FORZADO.equals(estado.getEstadoPublicacion().getCodigo()))){
 					// Si el estado anterior es despublicado y el actual es publicado, se reinicia el contador de días.
@@ -1136,11 +1202,7 @@ public class ActivoManager extends BusinessOperationOverrider<ActivoApi> impleme
 		DtoDatosPublicacion dto = new DtoDatosPublicacion();
 		dto.setIdActivo(idActivo);
 		dto.setTotalDiasPublicado(dias);
-		if(estadoPublicacionPortalesExternos){
-			dto.setPortalesExternos(this.ESTADO_PORTALES_EXTERNOS_PUBLICADO);
-		} else {
-			dto.setPortalesExternos(this.ESTADO_PORTALES_EXTERNOS_NO_PUBLICADO);
-		}
+		
 		return dto;
 	}
 
@@ -1251,6 +1313,12 @@ public class ActivoManager extends BusinessOperationOverrider<ActivoApi> impleme
 		if(Checks.esNulo(perimetroActivo)){
 			perimetroActivo = new PerimetroActivo();
 			perimetroActivo.setAuditoria(new Auditoria());
+			//Si no existia perimetro en BBDD, se deben tomar todas las condiciones marcadas
+			perimetroActivo.setAplicaTramiteAdmision(1);
+			perimetroActivo.setAplicaGestion(1);
+			perimetroActivo.setAplicaAsignarMediador(1);
+			perimetroActivo.setAplicaComercializar(1);
+			perimetroActivo.setAplicaFormalizar(1);
 		}
 		
 		return perimetroActivo;
@@ -1295,6 +1363,47 @@ public class ActivoManager extends BusinessOperationOverrider<ActivoApi> impleme
 			logger.error(ex.getMessage());
 			ex.printStackTrace();
 		}
+		
+		return perimetroActivo;
+	}
+	
+	@Override
+	@Transactional(readOnly = false)
+	public Activo updateActivoAsistida(Activo activo){
+		//Actualizamos el perímetro
+		PerimetroActivo perimetro = getPerimetroByIdActivo(activo.getId());
+		
+		if(Checks.esNulo(perimetro.getActivo()))
+			perimetro.setActivo(activo);
+		
+		updatePerimetroAsistida(perimetro);
+		
+		//Bloqueamos los precios para que el activo no salga en los procesos automáticos. Esto podría ir en un proceso al dar de alta el activo.
+		activo.setBloqueoPrecioFechaIni(new Date());
+		activo.setGestorBloqueoPrecio(adapter.getUsuarioLogado());
+		
+		saveOrUpdate(activo);
+		return activo;
+	}
+	
+	@Override
+	@Transactional(readOnly = false)
+	public PerimetroActivo updatePerimetroAsistida(PerimetroActivo perimetroActivo){
+		perimetroActivo.setIncluidoEnPerimetro(1);
+		perimetroActivo.setAplicaAsignarMediador(0);
+		perimetroActivo.setAplicaComercializar(1);
+		perimetroActivo.setAplicaFormalizar(1);
+		perimetroActivo.setAplicaGestion(0);
+		perimetroActivo.setAplicaTramiteAdmision(0);
+		perimetroActivo.setFechaAplicaComercializar(new Date());
+		perimetroActivo.setFechaAplicaFormalizar(new Date());
+		
+		
+		Filter filtro = genericDao.createFilter(FilterType.EQUALS, "codigo", DDMotivoComercializacion.CODIGO_ASISTIDA);
+		DDMotivoComercializacion motivoComercializacion = genericDao.get(DDMotivoComercializacion.class, filtro);
+		perimetroActivo.setMotivoAplicaComercializar(motivoComercializacion);
+		
+		saveOrUpdatePerimetroActivo(perimetroActivo);
 		
 		return perimetroActivo;
 	}
@@ -1447,4 +1556,110 @@ public class ActivoManager extends BusinessOperationOverrider<ActivoApi> impleme
 		return true;
 	}
 	
+	@Override
+	public boolean isIntegradoAgrupacionAsistida(Activo activo) {
+
+		for(ActivoAgrupacionActivo agrupacionActivo : activo.getAgrupaciones()) {
+
+			Date fechaFinVigencia = agrupacionActivo.getAgrupacion().getFechaFinVigencia();
+			fechaFinVigencia = !Checks.esNulo(fechaFinVigencia) ? new Date(fechaFinVigencia.getTime()) : null;
+			
+			if(!Checks.esNulo(agrupacionActivo.getAgrupacion().getTipoAgrupacion()) && 
+				agrupacionActivo.getAgrupacion().getTipoAgrupacion().getCodigo().equals(DDTipoAgrupacion.AGRUPACION_ASISTIDA) &&
+				!Checks.esNulo(fechaFinVigencia) && fechaFinVigencia.after(new Date())) 
+			{
+				return true;
+			}
+		}
+		
+		return false;
+	}
+	
+	@Override
+	public boolean isActivoAsistido(Activo activo){
+		if(!Checks.esNulo(activo.getSubcartera()))
+			if(DDSubcartera.CODIGO_CAJ_ASISTIDA.equals(activo.getSubcartera().getCodigo())
+					|| DDSubcartera.CODIGO_SAR_ASISTIDA.equals(activo.getSubcartera().getCodigo())
+					|| DDSubcartera.CODIGO_BAN_ASISTIDA.equals(activo.getSubcartera().getCodigo()))
+						return true;
+		return false;
+	}
+
+	@Override
+	public Integer getNumActivosPublicadosByAgrupacion(List<ActivoAgrupacionActivo> activos) {
+		
+		Integer contador = 0;
+		
+		for(ActivoAgrupacionActivo activoAgrupacion : activos) {
+			if(!Checks.esNulo(activoAgrupacion.getActivo().getEstadoPublicacion())) {
+				
+				String codEstadoPublicacion = activoAgrupacion.getActivo().getEstadoPublicacion().getCodigo();
+				if(codEstadoPublicacion.equals(DDEstadoPublicacion.CODIGO_PUBLICADO) || codEstadoPublicacion.equals(DDEstadoPublicacion.CODIGO_PUBLICADO_FORZADO) ||
+					codEstadoPublicacion.equals(DDEstadoPublicacion.CODIGO_PUBLICADO_PRECIOOCULTO) || codEstadoPublicacion.equals(DDEstadoPublicacion.CODIGO_PUBLICADO_OCULTO) ||
+					codEstadoPublicacion.equals(DDEstadoPublicacion.CODIGO_PUBLICADO_FORZADO_PRECIOOCULTO)) 
+				{
+					contador++;
+				}
+			}	
+		}
+		
+		return contador;
+	}
+
+	@Override
+	@Transactional(readOnly = false)
+	public Boolean solicitarTasacion(Long idActivo) {
+		
+		try{
+			// Se especifica bankia por que tan solo se va a poder demandar la tasación desde bankia.
+			int tasacionID = uvemManagerApi.ejecutarSolicitarTasacion(idActivo, adapter.getUsuarioLogado().getNombre(), "BANKIA");
+			if(!Checks.esNulo(tasacionID)){
+				Activo activo = activoDao.get(idActivo);
+				
+				if(!Checks.esNulo(activo)) {
+					// Generar un 'BIE_VALORACION' con el 'BIEN_ID' del activo.
+					NMBValoracionesBien valoracionBien = new NMBValoracionesBien();
+					beanUtilNotNull.copyProperty(valoracionBien, "bien", activo.getBien());
+					valoracionBien = genericDao.save(NMBValoracionesBien.class, valoracionBien);
+					
+					if(!Checks.esNulo(valoracionBien)) {
+						// Generar una tasacion con el ID de activo y el ID de la valoracion del bien.
+						ActivoTasacion tasacion = new ActivoTasacion();
+						
+						beanUtilNotNull.copyProperty(tasacion, "idExterno", tasacionID);
+						beanUtilNotNull.copyProperty(tasacion, "activo", activo);
+						beanUtilNotNull.copyProperty(tasacion, "valoracionBien", valoracionBien);
+						
+						genericDao.save(ActivoTasacion.class, tasacion);
+					}
+				}
+			}
+		}catch(Exception e){
+			e.printStackTrace();
+			return false;
+		}
+		
+		return true;
+	}
+
+	@Override
+	public DtoTasacion getSolicitudTasacionBankia(Long id) {
+		ActivoTasacion activoTasacion = activoDao.getActivoTasacion(id);
+		DtoTasacion dtoTasacion = new DtoTasacion();
+		if(!Checks.esNulo(activoTasacion)) {
+			
+			try {
+				beanUtilNotNull.copyProperty(dtoTasacion, "id", activoTasacion.getId());
+				beanUtilNotNull.copyProperty(dtoTasacion, "fechaSolicitudTasacion", activoTasacion.getAuditoria().getFechaCrear());
+				beanUtilNotNull.copyProperty(dtoTasacion, "gestorSolicitud", activoTasacion.getAuditoria().getUsuarioCrear());
+				beanUtilNotNull.copyProperty(dtoTasacion, "externoID", activoTasacion.getIdExterno());
+			} catch (IllegalAccessException e) {
+				e.printStackTrace();
+			} catch (InvocationTargetException e) {
+				e.printStackTrace();
+			}
+		}
+		
+		return dtoTasacion;
+	}
 }
